rename `cargo-sqlx` -> `sqlx-cli`
edit README
