+++ sqlx-cli/src/bin/sqlx.rs
@@ -1,8 +1,8 @@
-use cargo_sqlx::Command;
+use sqlx_cli::Command;
 use structopt::StructOpt;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
     // no special handling here
-    cargo_sqlx::run(Command::from_args()).await
+    sqlx_cli::run(Command::from_args()).await
 }
